+++ app/src/main/java/edu/cascadia/mobas/pollinatorpathway/Database/Planting/Planting.java
@@ -2,6 +2,7 @@ package edu.cascadia.mobas.pollinatorpathway.Database.Planting;
 
 import static androidx.room.ForeignKey.CASCADE;
 
+import androidx.annotation.NonNull;
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.ForeignKey;
@@ -18,37 +19,19 @@ public class Planting {
     @ColumnInfo(name = "plantingId")
     private long plantingId;
 
-    @ColumnInfo(name = "address")
-    String address;
-
     @ColumnInfo(name = "name")
     String name;
 
     @ColumnInfo(name = "description")
     String description;
 
-    @ColumnInfo(name = "type")
-    String type;
-
-    @ColumnInfo(name = "website")
-    String website;
-
-    @ColumnInfo(name = "dateJoined")
-    String dateJoined;
 
     @ColumnInfo(name = "gpsLocation")
     String gpsLocation;
 
+
     @ColumnInfo(name = "userId")
     private long userId;
-
-
-
-
-
-
-
-
 
 
     public long getPlantingId() {
@@ -67,14 +50,6 @@ public class Planting {
         this.name = name;
     }
 
-    public String getType() { return type; }
-
-    public void setType(String type) { this.type = type; }
-
-    public String getDateJoined() { return dateJoined; }
-
-    public void setDateJoined(String dateJoined) { this.dateJoined = dateJoined; }
-
     public String getDescription() {
         return description;
     }
@@ -83,21 +58,11 @@ public class Planting {
         this.description = description;
     }
 
-    public String getWebsite() { return website; }
-
-    public void setWebsite(String website) { this.website = website; }
 
     public String getGpsLocation() { return gpsLocation; }
 
     public void setGpsLocation(String gpsLocation) { this.gpsLocation = gpsLocation; }
 
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
 
     public long getUserId() {
         return userId;
